+++ CHANGELOG.md
@@ -21,6 +21,8 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * Made the content filter mechanism more flexible, so that filters can be made to run only on specific events.
 * Made Ajax hooks run immediately by default to minimize flicker; let's see if this breaks anything...
 * Content filters are now applied to answers loaded in review by the [mse172931](http://meta.stackexchange.com/q/172931) fix.
+* New utility function `SOUP.hookChat()` for monitoring chat events.
+* Chat content filters now use the page visibility API to only run in active tabs.
 
 1.24 (7 Oct 2014)
 ====